refactor(rust): gate Discord presence helpers
